Improve filesystem check, improve wording
I made it so that the script checks that the data partition is either F2FS or EXT4, not if it is not vFAT. 

However, I tried to improve it to make more sense, and direct people to post 2. 

Because it will hopefully further prevent [this](http://forum.xda-developers.com/showpost.php?p=68981820&postcount=1093). -_-